requirements.txt
Django==3.0.2
django-cors-headers==3.2.1
django-extensions==2.2.6
djangorestframework==3.11.0
mysqlclient==2.0.1
django-mysql==3.8.1
ipdb==0.12.3
ipython==7.12.0